core changes for refactor including pyproject.toml updates and renaming config.py to settings.py due to later config additions and potential conflicts in naming
