+++ src/guidellm/__main__.py
@@ -13,9 +13,9 @@ from guidellm.benchmark import (
 )
 from guidellm.benchmark.entrypoints import benchmark_with_scenario
 from guidellm.benchmark.scenario import GenerativeTextScenario, get_builtin_scenarios
-from guidellm.config import print_config
 from guidellm.preprocess.dataset import ShortPromptStrategy, process_dataset
 from guidellm.scheduler import StrategyType
+from guidellm.settings import print_config
 from guidellm.utils import DefaultGroupHandler
 from guidellm.utils import cli as cli_tools
