+++ src/guidellm/backend/openai.py
@@ -16,7 +16,7 @@ from guidellm.backend.response import (
     ResponseSummary,
     StreamingTextResponse,
 )
-from guidellm.config import settings
+from guidellm.settings import settings
 
 __all__ = [
     "CHAT_COMPLETIONS",
