+++ src/guidellm/backend/response.py
@@ -2,8 +2,8 @@ from typing import Any, Literal, Optional
 
 from pydantic import computed_field
 
-from guidellm.config import settings
 from guidellm.objects.pydantic import StandardBaseModel
+from guidellm.settings import settings
 
 __all__ = [
     "RequestArgs",
