+++ src/guidellm/scheduler/strategy.py
@@ -10,8 +10,8 @@ from typing import (
 
 from pydantic import Field
 
-from guidellm.config import settings
 from guidellm.objects import StandardBaseModel
+from guidellm.settings import settings
 
 __all__ = [
     "AsyncConstantStrategy",
