+++ src/guidellm/utils/text.py
@@ -10,7 +10,7 @@ import httpx
 from loguru import logger
 
 from guidellm import data as package_data
-from guidellm.config import settings
+from guidellm.settings import settings
 
 __all__ = [
     "EndlessTextCreator",
