+++ tests/unit/backend/test_openai_backend.py
@@ -3,7 +3,7 @@ import time
 import pytest
 
 from guidellm.backend import OpenAIHTTPBackend, ResponseSummary, StreamingTextResponse
-from guidellm.config import settings
+from guidellm.settings import settings
 
 
 @pytest.mark.smoke
